Completed refactor incubator flow
