+++ routes/routes.js
@@ -1,0 +1,13 @@
+import CoopRoutes from './coopRoutes.js';
+import ChickenRoutes from './chickenRoutes.js';
+import BrooderRoutes from './brooderRoutes.js';
+import IncubatorRoutes from './incubatorRoutes.js';
+
+// The purpose of this file is to gather all different routes so that app.js need only 1 import route
+
+export {
+  CoopRoutes,
+  ChickenRoutes,
+  BrooderRoutes,
+  IncubatorRoutes
+};
