+++ views/create-hatch-record.ejs
@@ -151,7 +151,7 @@
     %>
 
     <div class="p-3 mx-5 border border-secondary rounded shadow-lg mb-4">
-      <form class="needs-validation" action="/submit-hatch-record" method="post" role="form" class="text-start">
+      <form class="needs-validation" action="/incubator/submit-hatch-record" method="post" role="form" class="text-start">
         <div id="formRepeat">
           <div class="row border px-2 py-2 mx-1">
             <div class="row mt-1">
