+++ views/create-tray-record.ejs
@@ -160,7 +160,7 @@
     %>
 
     <div class="p-3 mx-5 border border-secondary rounded shadow-lg mb-4">
-      <form class="needs-validation" action="/submit-tray-record" method="post" role="form" class="text-start">
+      <form class="needs-validation" action="/incubator/submit-tray-record" method="post" role="form" class="text-start">
         <div id="formRepeat">
           <div class="row border px-2 py-2 mx-1">
             <div class="row mt-1">
